Played around a little
Draws an octahedron for lighting testing. Bad news is, that I found out,
that having the lighting how I wanted it, requires much math, and would
therefore be more impacting on the performance, than is acceptable. Good
news is, that nevertheless you can now pretty much see all the shapes,
so maybe we can leave it that way.